% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download_mortality.R
\name{download_mortality}
\alias{download_mortality}
\title{Import Death Data}
\usage{
download_mortality(year)
}
\arguments{
\item{year}{Tde year of data to download}
}
\description{
Import Death Data from the web
}
\details{
This will load data from web, may grab one or two files depending on the year. The return is simply the name of the file
}
\examples{
download_mortality(2007)
}

