% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import.R
\name{import}
\alias{import}
\title{Import from file}
\usage{
import(file)
}
\arguments{
\item{file}{An unzipped file from the cdc website}
}
\description{
Read birth data from file
}
\details{
Load data from fixed width file to R data.frame
}
\examples{
import('Nat2013PublicPS.r20141016')
}

