% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/import_birth.R
\name{import_birth}
\alias{import_birth}
\title{Import Birth Data}
\usage{
import_birth(set)
}
\arguments{
\item{set}{first or section portion}
}
\description{
Import Birth Data from the web
}
\details{
This will load data from web
}
\examples{
import_birth()
}

